+++ appmap/test/conftest.py
@@ -1,8 +1,10 @@
+from distutils.dir_util import copy_tree
 import importlib
 import pytest
 import yaml
 
 import appmap._implementation
+from appmap._implementation import utils
 from appmap._implementation.env import Env
 from appmap._implementation.recording import Recorder
 
@@ -43,3 +45,24 @@ def pytest_runtest_setup(item):
     # Some tests want yaml instrumented, others don't.
     # Reload it to make sure it's instrumented, or not, as set in appmap.yml.
     importlib.reload(yaml)
+
+@pytest.fixture(scope='session', name='git_directory')
+def git_directory_fixture(tmp_path_factory):
+    git_dir = tmp_path_factory.mktemp('test_project')
+    (git_dir / 'README.metadata').write_text('Read me')
+    (git_dir / 'new_file').write_text('new_file')
+
+    git = utils.git(cwd=git_dir)
+    git('init')
+    git('config --local user.email test@test')
+    git('config --local user.name Test')
+    git('add README.metadata')
+    git('commit -m "initial commit"')
+    git('remote add origin https://www.example.test/repo.git')
+
+    return git_dir
+
+@pytest.fixture(name='git')
+def tmp_git(git_directory, tmp_path):
+    copy_tree(git_directory, str(tmp_path))
+    return utils.git(cwd=tmp_path)