+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [1.2.0.dev1] - 2021-07-16
 ### Added
 - The `appmap-agent-init` and `appmap-agent-status` internal commands are now available
   for the code-editor extensions.
@@ -272,7 +273,8 @@ testing. Also adds a Travis config to run them all.
 - Deploy release to PyPI
 
 
-[Unreleased]: https://github.com/applandinc/appmap-python/compare/1.1.0...HEAD
+[Unreleased]: https://github.com/applandinc/appmap-python/compare/1.2.0.dev1...HEAD
+[1.2.0.dev1]: https://github.com/applandinc/appmap-python/compare/1.1.0...1.2.0.dev1
 [1.1.0]: https://github.com/applandinc/appmap-python/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/applandinc/appmap-python/compare/0.10.0...1.0.0
 [0.10.0]: https://github.com/applandinc/appmap-python/compare/0.9.0...0.10.0